% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_comparison.R
\name{compare_column_elements}
\alias{compare_column_elements}
\title{Compare Elements Between 2 Columns}
\usage{
compare_column_elements(list1, list2)
}
\arguments{
\item{list1}{first column to compare}

\item{list2}{second column to compare}
}
\description{
If both columns contain the same elements, a message is displayed
If the columns do not contain the same elements, 3 elements are printed: elements on list A but not on list B, elements on list B but not on list A and common elements
}
