% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_clipboard.R
\name{copy_to_excel}
\alias{copy_to_excel}
\title{Copy to Excel (via clipboard)}
\usage{
copy_to_excel(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{dataframe to be copied to Excel}

\item{row.names}{Binary - Indicates whether the row names should be copied}

\item{col.names}{Binary - Indicates whether the column names should be copied}

\item{...}{params used in write.table function (from utils package)}
}
\value{
None
}
\description{
Copy to Excel (via clipboard)
}
