% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_clipboard.R
\name{paste_from_excel}
\alias{paste_from_excel}
\title{Paste from Excel (via clipboard)}
\usage{
paste_from_excel(header = TRUE, ...)
}
\arguments{
\item{header}{indicates wether the copied region contains a header row}

\item{...}{params used in read.table function (from utils package)}
}
\value{
None
}
\description{
Imports a copied Excel region into a dataframe
}
