% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_revi.R
\name{theme_revi}
\alias{theme_revi}
\title{Revi theme, based on Theme Ipsum RC by Hrbrthemes}
\usage{
theme_revi(border = TRUE, ...)
}
\arguments{
\item{border}{Indicates whether the panel border should be included on the plot}

\item{...}{params for theme_ipsum_rc}
}
\description{
Revi theme, based on Theme Ipsum RC by Hrbrthemes
}
